Update on "[1/n] Use uint16_t for MesssageType enum"
Context: to support custom messages, the `MessageType` field in RPC
Message can no longer be an enum. We plan to change that to an
`unint16_t` and allow users to register their own types.

This commit explicilty uses `uint16_t` as the integer type for the
enum.

cc pietern mrshenli pritamdamania87 zhaojuanmao satgera rohan-varma gqchen aazzolini osalpekar jiayisuse @SciPioneer @H-Huang

Differential Revision: [D32263167](https://our.internmc.facebook.com/intern/diff/D32263167)

[ghstack-poisoned]

+++ torch/csrc/distributed/autograd/rpc_messages/rpc_with_profiling_req.cpp
@@ -27,7 +27,7 @@ RpcWithProfilingReq::RpcWithProfilingReq(
       messageType_ == rpc::MessageType::RUN_WITH_PROFILING_REQ,
       c10::str(
           "Incorrect message type, expected message type ",
-          uint16_t(rpc::MessageType::RUN_WITH_PROFILING_REQ)));
+          rpc::MessageType::RUN_WITH_PROFILING_REQ));
   wrappedMessageType_ = wrappedMessage_->type();
 }
